+++ koishi/plugins/automation_logging/__init__.py
@@ -1,4 +1,5 @@
 from .components_satori_auto_start import *
+from .constants import *
 from .embed_builder_emoji import *
 from .embed_builder_guild_profile import *
 from .embed_builder_mention import *
@@ -17,6 +18,7 @@ from .interactions_satori import *
 
 __all__ = (
     *components_satori_auto_start.__all__,
+    *constants.__all__,
     *embed_builder_emoji.__all__,
     *embed_builder_guild_profile.__all__,
     *embed_builder_mention.__all__,