+++ koishi/plugins/automation_logging/constants.py
@@ -1,0 +1,13 @@
+__all__ = ()
+
+from hata import Color
+
+
+COLOR_ADD = Color.from_rgb(2, 168, 77)
+COLOR_UPDATE = Color.from_rgb(237, 206, 28)
+COLOR_DELETE = Color.from_rgb(168, 49, 2)
+
+COLOR_SATORI = Color.from_rgb(168, 2, 146)
+
+IMAGE_URL_SATORI = 'https://cdn.discordapp.com/attachments/568837922288173058/1109762751963865200/satori-0015-edit-0000.png'
+