+++ koishi/plugins/automation_welcome/__init__.py
@@ -4,6 +4,7 @@ from .constants import *
 from .events import *
 from .helpers import *
 from .interactions import *
+from .reply_style import *
 from .spam_protection import *
 from .welcome_style import *
 
@@ -15,6 +16,7 @@ __all__ = (
     *events.__all__,
     *helpers.__all__,
     *interactions.__all__,
+    *reply_style.__all__,
     *spam_protection.__all__,
     *welcome_style.__all__,
 )