+++ koishi/plugins/automation_welcome/welcome_styles/koishi.py
@@ -2,6 +2,7 @@ __all__ = ()
 
 from hata import Emoji
 
+from ..reply_style import ReplyStyle
 from ..welcome_style import WelcomeStyle
 
 
@@ -17,7 +18,7 @@ MESSAGE_CONTENT_BUILDERS = (
     (lambda mention: f'Heart throbbing {mention} showed up!'),
     (lambda mention: f'It\'s a youkai! It\'s a fairy! Nevermind, it\'s just {mention}.'),
     (lambda mention: f'We\'ve been already behind you {mention}'),
-    (lambda mention: f'It\'s {mention}! Praise Mr. Unconsciousness! [T]/'),
+    (lambda mention: f'It\'s {mention}! Praise Mr. Unconsciousness! \\\\[T]/'),
     (lambda mention: f'Always gonna love {mention}. Always gonna remember {mention}.'),
     (lambda mention: f'{mention} is here to stab and eat shrimp fry. And {mention} is all out of shrimp fry.'),
     (lambda mention: f'{mention} is behind me... or are they?'),
@@ -371,46 +372,93 @@ IMAGES = (
     'https://cdn.discordapp.com/attachments/568837922288173058/1151772310215348244/koishi-welcome-0318.png',
 )
 
-BUTTON_EMOJI = Emoji.precreate(648173118392762449)
 
-BUTTON_CONTENTS = (
-    'Share mr. hat with them.',
-    'Share shrimp fry with them.',
-    'Help them embrace mr. unconsciousness.',
-    'Share your fishing rod with them.',
-    'Share croissant with them.',
-    'Share shrimp fry and croissant with them.',
-    'Hug them.',
-    'Share subconscious love with them.',
-    'Help them stop thinking.',
-    'Share koishi fumo with them.',
-    'To staaaaare at them.',
-    'To go fishing with them.',
-    'Welcome them to the Extra Stage.',
-    'Welcome them to a heart throbbing adventure.',
-    'Become their imaginary friend.',
-    'Call them.',
-    'Welcome them to the Palace of the Earth Spirits.',
-)
-
-REPLY_CONTENT_BUILDERS = (
-    (lambda source, target: f'{source} shares mr. hat with {target}.'),
-    (lambda source, target: f'{source} shares shrimp fry with {target}.'),
-    (lambda source, target: f'{source} helps {target} embrace mr. unconsciousness.'),
-    (lambda source, target: f'{source} shares their fishing rod with {target}.'),
-    (lambda source, target: f'{source} shares croissant with {target}.'),
-    (lambda source, target: f'{source} shares shrimp fry and croissant with {target}.'),
-    (lambda source, target: f'{source} hugs {target}.'),
-    (lambda source, target: f'{source} shares subconscious love with {target}.'),
-    (lambda source, target: f'{source} helps {target} stop thinking.'),
-    (lambda source, target: f'{source} shares koishi fumo with {target}.'),
-    (lambda source, target: f'{source} staaaaares at {target}.'),
-    (lambda source, target: f'{source} goes fishing with {target}.'),
-    (lambda source, target: f'{source} welcomes {target} to the Extra Stage.'),
-    (lambda source, target: f'{source} welcomes {target} to a heart throbbing adventure.'),
-    (lambda source, target: f'{source} becomes {target}\'s imaginary friend.'),
-    (lambda source, target: f'{source} calls {target}.'),
-    (lambda source, target: f'{source} welcomes {target} to the Palace of the Earth Spirits.'),
+REPLY_STYLES = (
+    ReplyStyle(
+        'Share mr. hat with them.',
+        Emoji.precreate(1175509886902218874),
+        (lambda source, target: f'{source} shares mr. hat with {target}.'),
+    ),
+    ReplyStyle(
+        'Share shrimp fry with them.',
+        Emoji.precreate(1175513087156166821),
+        (lambda source, target: f'{source} shares shrimp fry with {target}.'),
+    ),
+    ReplyStyle(
+        'Help them embrace mr. unconsciousness.',
+        Emoji.precreate(828622013798547466),
+        (lambda source, target: f'{source} helps {target} embrace mr. unconsciousness.'),
+    ),
+    ReplyStyle(
+        'Share your fishing rod with them.',
+        Emoji.precreate(1175508286267728053),
+        (lambda source, target: f'{source} shares their fishing rod with {target}.'),
+    ),
+    ReplyStyle(
+        'Share croissant with them.',
+        Emoji.precreate(1175513903518711838),
+        (lambda source, target: f'{source} shares croissant with {target}.'),
+    ),
+    ReplyStyle(
+        'Share shrimp fry and croissant with them.',
+        Emoji.precreate(1052193063222661151),
+        (lambda source, target: f'{source} shares shrimp fry and croissant with {target}.'),
+    ),
+    ReplyStyle(
+        'Hug them.',
+        Emoji.precreate(1148897323414454323),
+        (lambda source, target: f'{source} hugs {target}.'),
+    ),
+    ReplyStyle(
+        'Share subconscious love with them.',
+        Emoji.precreate(852856143528853514),
+        (lambda source, target: f'{source} shares subconscious love with {target}.'),
+    ),
+    ReplyStyle(
+        'Help them stop thinking.',
+        Emoji.precreate(813390875408203787),
+        (lambda source, target: f'{source} helps {target} stop thinking.'),
+    ),
+    ReplyStyle(
+        'Share koishi fumo with them.',
+        Emoji.precreate(1175520338793201816),
+        (lambda source, target: f'{source} shares koishi fumo with {target}.'),
+    ),
+    ReplyStyle(
+        'To staaaaare at them.',
+        Emoji.precreate(1083349493346480140),
+        (lambda source, target: f'{source} staaaaares at {target}.'),
+    ),
+    ReplyStyle(
+        'Go fishing with them.',
+        Emoji.precreate(1056922454884364318),
+        (lambda source, target: f'{source} goes fishing with {target}.'),
+    ),
+    ReplyStyle(
+        'Welcome them to the Extra Stage.',
+        Emoji.precreate(1048978997838614618),
+        (lambda source, target: f'{source} welcomes {target} to the Extra Stage.'),
+    ),
+    ReplyStyle(
+        'Welcome them to a heart throbbing adventure.',
+        Emoji.precreate(1029100558264586250),
+        (lambda source, target: f'{source} welcomes {target} to a heart throbbing adventure.'),
+    ),
+    ReplyStyle(
+        'Become their imaginary friend.',
+        Emoji.precreate(1029100497526865920),
+        (lambda source, target: f'{source} becomes {target}\'s imaginary friend.'),
+    ),
+    ReplyStyle(
+        'Call them.',
+        Emoji.precreate(1175506548022321262),
+        (lambda source, target: f'{source} calls {target}.'),
+    ),
+    ReplyStyle(
+        'Welcome them to the Palace of the Earth Spirits.',
+        Emoji.precreate(1031941507881644093),
+        (lambda source, target: f'{source} welcomes {target} to the Palace of the Earth Spirits.'),
+    ),
 )
 
 
@@ -418,7 +466,5 @@ WELCOME_STYLE = WelcomeStyle(
     NAME,
     MESSAGE_CONTENT_BUILDERS,
     IMAGES,
-    BUTTON_EMOJI,
-    BUTTON_CONTENTS,
-    REPLY_CONTENT_BUILDERS,
+    REPLY_STYLES,
 )